Move BerserkersTorment normalizers to the other normalizers
